feat:Add authorization function

+++ controllers/routers.py
@@ -3,12 +3,13 @@ from sqlalchemy.orm import Session
 from database.database import get_db
 from database.schemas import RequestUser, Response
 from controllers import crud
-from security.middleware_check import check_jwt_token
+from security.middleware_check import check_jwt_token,check_admin
 
 router = APIRouter()
 #Đường dẫn thêm user
 @router.post('/create')
 async def create_user(request: RequestUser, db: Session = Depends(get_db), token_payload: dict = Depends(check_jwt_token)):
+    await check_admin(token_payload)
     created_user = crud.create_user(db, request)
     return Response(user=created_user).dict(exclude_none=True)
 #Đường dẫn lấy tất cả user
@@ -29,5 +30,6 @@ async def update_user(request: RequestUser, id: int, db: Session = Depends(get_d
 #Đường dẫn xóa user theo id
 @router.delete('/{id}')
 async def delete_user(id: int, db: Session = Depends(get_db), token_payload: dict = Depends(check_jwt_token)):
+    await check_admin(token_payload)
     deleted = crud.delete_user(db, id)
     return {"success": deleted}
